Rename light/dark icon components
